Move gdrive download to util.py
